CHG: Version number is now version 1.14. FIX: If we get two speed messages very close together, it could happen we would be one message delayed on each message. This could re-happen also. Now this should not happen. UPD: First version compiled under Delphi 10.2 Tokyo. UPD: The executables have been updated in the release.

+++ Callidus/CallidusController/fCallidusController.pas
@@ -442,7 +442,7 @@ begin
     iSecondeSansService := ((GetTickCount - FreezeLastServiceReceived) div 1000);
     MyStatusBar.Panels[IDX_PANEL_LASTSERVICE].Text := Format('%d min %2.2d sec sans service', [(iSecondeSansService div 60), (iSecondeSansService - ((iSecondeSansService div 60) * 60))]);
 
-    if (ckbControlePubAlone.Checked) and (iSecondeSansService > ((cbInactivitePeriodStartPub.ItemIndex + 1) * 60)) and (not bModePublicite) then
+    if (ckbControlePubAlone.Checked) and (iSecondeSansService > DWORD((cbInactivitePeriodStartPub.ItemIndex + 1) * 60)) and (not bModePublicite) then
       actStartPubExecute(actStartPub);
   end;
 end;
